refactor document service
